requirements.txt
alembic
aniso8601
attrs
blinker
click
colorama
dnspython
email-validator
Flask
Flask-Login
Flask-Migrate
Flask-Minify
flask-restx
Flask-WTF
greenlet
gunicorn
htmlmin
idna
itsdangerous
Jinja2
jsmin
jsonschema
jsonschema-specifications
lesscpy
Mako
MarkupSafe
pip
ply
python-dotenv
pytz
rcssmin
referencing
rpds-py
setuptools
six
typing_extensions
Werkzeug
WTForms
xxhash
# flask_mysqldb
